requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.4.0
opencv-python>=4.7.0
scikit-image>=0.21.0

# Transformers for ViT
timm>=0.9.0

# Metrics
pytorch-fid>=0.3.0
lpips>=0.1.4
torchmetrics>=1.0.0

# Data handling
pandas>=2.0.0
kaggle>=1.5.0
tqdm>=4.65.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
tensorboard>=2.13.0

# Web UI
gradio>=4.0.0

# Config
PyYAML>=6.0
omegaconf>=2.3.0

# Stain normalization
staintools>=2.1.2
